requirements.txt
aiohttp==3.9.3
b2sdk==2.8.0
beautifulsoup4==4.12.3
diskcache==5.6.3
fastapi==0.109.2
html2text==2020.1.16
httpx==0.26.0
ipykernel==6.29.3
jinaai==0.2.10
meilisearch==0.34.1
neo4j==5.20.0
notebook==7.1.2
numpy==1.26.3
pandas==2.2.2
pydantic==2.6.1
pydantic-settings==2.2.1
pytest==8.3.4
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
qdrant-client==1.7.0
requests==2.31.0
requests-cache==1.1.1
requests-mock==1.11.0
responses==0.25.0
scikit-learn==1.4.0
smolagents==0.1.0
torch==2.2.0
tqdm==4.66.1
transformers==4.37.2
uvicorn==0.27.1
pyalex
springernature-api-client
